Reflection
What went well?
Connecting the dots between the examples that we did in class and this exercise. I thought that the class examples were a really great reflection of what the assignment was and that we went through it piece by piece so that it didn't feel overwhelming. Nice job, instructors! 
It should be said that the image was only inserted as a revenge tactic for our instructors 😈 /method of making me and my peers laugh 😂 /encouragement for the user😁 . It was only awkward when I forgot about it and asked a tutor for help...😑 😂

What are you still working on understanding better?
- I'd like to gain further understanding of the different buttons available. I used a checkbox when creating a new task if the user wanted to mark the task as complete, which does work. However, if you go back and edit a completed task in the edit view it will show that checkbox as unchecked which makes it look like it is actually not completed. Or, if you forget to re-check it when you submit the edits it actually marks the task as incomplete, which wouldn't be accurate. My chair pair told me that she talked about this with Chris and he mentioned that this would require some Javascript so maybe this is something I would be able to fix in the future, but not now.

What did your chair pair do differently?
My chair pair has a layout that looks very list-y. She even added a counter that lets you know how many things you've completed as well as suggests things that are still incomplete on your list. I thought that was a great addition, since it's a great reminder of things still to do!